pyproject.toml
[project]
name = "vecs"
version = "0.1.0"
description = "Semantic search for your codebase and Claude Code sessions"
requires-python = ">=3.12"
dependencies = [
    "voyageai>=0.3",
    "chromadb>=1.0",
    "click>=8.0",
    "mcp[cli]>=1.2",
    "tree-sitter>=0.24",
    "tree-sitter-c-sharp>=0.23",
    "tree-sitter-typescript>=0.23",
    "rank-bm25>=0.2",
    "cachetools>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
vecs = "vecs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
